fix: changed double to integer for Height

+++ src/main/java/com/alessandropesole/bonwoapp/user/application/dto/UserProfileResponse.java
@@ -11,7 +11,7 @@ public record UserProfileResponse(
         ImageResponse avatar,
         String bio,
         Integer ageYears,
-        Double heightCm,
+        Integer heightCm,
         Double weightKg,
         List<ActivityResponse> activities
 ) {}
